+++ src/other/proppedModel.ts
@@ -1,6 +1,7 @@
 import * as fs from 'fs';
 import * as path from 'path';
 import { getBuffer } from 'gltf-import-export';
+import { Box, Quaternion, Vector, Vector2, sortVectorsByOutline } from './math';
 
 interface IProp {
   /* eslint-disable @typescript-eslint/naming-convention */
@@ -174,14 +175,20 @@ export default class ProppedModel {
   private readonly gltf: any;
   private readonly resourceFolder: string;
 
-  private groundVertices: { xf: number, zf: number }[] | undefined;
-
   public static fromFile(filePath: string) {
     const gltf = JSON.parse(fs.readFileSync(filePath, 'utf8'));
     const props: IPropMap = { };
     const particles: IParticleMap = { };
     const feedbacks: IFeedbackMap = { };
     const files: IFileMap = { };
+
+    // create empty arrays to reduce the need for undefined checks
+    if (!gltf.nodes) {
+      gltf.nodes = [];
+    }
+    if (!gltf.meshes) {
+      gltf.meshes = [];
+    }
     
     for (let node of gltf.nodes) {
       if (node.name.startsWith('prop_')) {
@@ -305,53 +312,87 @@ export default class ProppedModel {
 
   public getBuildBlocker() {
     const ground = this._findGround();
-    if (!ground) {
-      return;
+    if (!ground || ground.length === 0) {
+      return undefined;
     }
 
-    const result = [];
-    for (let vertex of ground) {
-      // round to .5
-      result.push({
-        xf: Math.round(vertex.xf * 2) / 2,
-        zf: Math.round(vertex.zf * 2) / 2
-      });
-    }
-
-    return result;
+    // round to .5
+    return ground.map(v => v.round(2));
   }
 
   public getDecalExtends() {
     const ground = this._findGround();
-    if (!ground) {
-      return;
+    if (!ground || ground.length < 1) {
+      return undefined;
     }
 
-    const result = [];
-
-    let minX = Math.round(ground[0].xf * 100) / 100;
-    let minZ = Math.round(ground[0].zf * 100) / 100;
-    let maxX = minX;
-    let maxZ = minZ;
-
-    for (let vertex of ground) {   
-      const xf = Math.round(vertex.xf * 100) / 100;
-      const zf = Math.round(vertex.zf * 100) / 100;
-      minX = Math.min(minX, xf);
-      minZ = Math.min(minZ, zf);
-      maxX = Math.max(maxX, xf);
-      maxZ = Math.max(maxZ, zf);
+    // round to .01
+    let min = ground[0].round(100);
+    let max = min;
+    for (let v of ground) {
+      min = min.down(v.round(100));
+      max = max.up(v.round(100));
     }
-
-    // TODO if not centered set Transformer, but that's unusual
 
     /* eslint-disable @typescript-eslint/naming-convention */
     return { 
-      Extents_x: ((maxX - minX) / 2).toFixed(6), 
+      Extents_x: ((max.x - min.x) / 2).toFixed(6), 
       Extents_y: (0.25).toFixed(6), 
-      Extents_z: ((maxZ - minZ) / 2).toFixed(6)
+      Extents_z: ((max.z - min.z) / 2).toFixed(6)
     };
     /* eslint-enable @typescript-eslint/naming-convention */
+  }
+
+  public getHitBoxes() {
+    return this._findHitboxes();
+  }
+
+  private _unevenBlocker: Vector2[] | undefined;
+  public getUnevenBlocker() {
+    if (!this._unevenBlocker) {
+      const node = _findFirstNode(this.gltf, 'UnevenBlocker', this.resourceFolder);
+      let vectors = _readVectors(node).map(e => e.toVector2());
+      if (node?.indices) {
+        vectors = sortVectorsByOutline(vectors, node.indices);
+      }
+      this._unevenBlocker = vectors;
+    }
+    return (this._unevenBlocker.length > 0) ? this._unevenBlocker : undefined;
+  }
+
+  private _feedbackBlocker: Vector2[][] | undefined;
+  public getFeedbackBlocker() {
+    if (!this._feedbackBlocker) {
+      this._feedbackBlocker = [];
+      const nodes = _findNodes(this.gltf, 'FeedbackBlocker', this.resourceFolder);
+      for (let node of nodes) {
+        let vectors = _readVectors(node).map(e => e.toVector2());
+        if (node?.indices) {
+          vectors = sortVectorsByOutline(vectors, node.indices);
+        }
+        if (vectors) {
+          this._feedbackBlocker.push(vectors);
+        }
+      }
+    }
+    return (this._feedbackBlocker.length > 0) ? this._feedbackBlocker : undefined;
+  }
+
+  private _dummies: { name: string, position: Vector, rotation: Quaternion, extends: Vector }[] | undefined;
+  public getDummies() {
+    if (!this._dummies) {
+      this._dummies = [];
+      const nodes = _findNodes(this.gltf, 'dummy_', this.resourceFolder);
+      for (let node of nodes) {
+        this._dummies.push({
+          name: node.name.substr('dummy_'.length),
+          position: node.translation || Vector.zero,
+          rotation: node.rotation || Quaternion.default,
+          extends: node.scale || Vector.one
+        });
+      }
+    }
+    return (this._dummies.length > 0) ? this._dummies : undefined;
   }
 
   private constructor(gltf: any, props: IPropMap, particles: IParticleMap, feedbacks: IFeedbackMap, files: IFileMap, resourceFolder: string) {
@@ -363,63 +404,172 @@ export default class ProppedModel {
     this.resourceFolder = resourceFolder;
   }
 
-  private _getBuffer(gltf: any, meshIdx: number, resourceFolder: string) {
-    const accessorIdx = gltf.meshes[meshIdx].primitives[0].attributes.POSITION;
-    const accessor = gltf.accessors[accessorIdx];
-    const bufferView = gltf.bufferViews[accessor.bufferView];
-    const bufferInfo = gltf.buffers[bufferView.buffer];
-
-    const bufferFile = path.join(resourceFolder, bufferInfo.uri);
-    const buffer = getBuffer(gltf, bufferView.buffer, bufferFile);
-    if (buffer) {
-      const bufferOffset = bufferView.byteOffset || 0;
-      const bufferLength = bufferView.byteLength;
-      const bufferStride = bufferView.byteStride;
-      const bufferViewBuf = buffer.subarray(bufferOffset, bufferOffset + bufferLength);
-      const accessorByteOffset = accessor.byteOffset || 0;
-
-      const ACESSOR_TYPE_VEC3 = 3;
-      return buildArrayBuffer(Float32Array, bufferViewBuf, accessorByteOffset, accessor.count, ACESSOR_TYPE_VEC3, bufferStride);
-    }
-
-    return undefined;
-  }
-
+  private groundVertices: Vector2[] | undefined;
   private _findGround() {
     if (!this.groundVertices) {
-      let groundMeshIdx = -1;
-      for (let node of this.gltf.nodes) {
-        if (node.name === 'ground' || this.gltf.meshes[node.mesh]?.name === 'ground') {
-          groundMeshIdx = node.mesh;
-        }
-      }
-
-      if (groundMeshIdx < 0) {
-        console.warn('Invalid glTF. Invalid mesh index.');
-        return undefined;
-      }
-
-      const buffer = this._getBuffer(this.gltf, groundMeshIdx, this.resourceFolder);
-      if (!buffer) {
-        console.warn('Invalid glTF. Could not get buffer.');
-        return undefined;
-      }
-
-      this.groundVertices = [];
-      for (let i = 0; i < 4 && i < buffer.length / 3; i++) {
-        // round to .5
-        this.groundVertices[i] = {
-          xf: buffer[i * 3],
-          zf: buffer[i * 3 + 2]
-        };
+      const node = _findFirstNode(this.gltf, 'ground', this.resourceFolder);
+      this.groundVertices = _readVectors(node).map(e => e.toVector2());
+      if (node?.indices) {
+        this.groundVertices = sortVectorsByOutline(this.groundVertices, node.indices);
       }
     }
 
     return this.groundVertices;
+  }
+
+  _hitboxes: Box[] | undefined;
+  private _findHitboxes() {
+    if (!this._hitboxes) {
+      this._hitboxes = [];
+      const hitboxes = [];
+      for (let node of this.gltf.nodes) {
+        if (node.name?.startsWith('hitbox')) {
+          hitboxes.push({ name: node.name, meshIdx: node.mesh, node: node });
+        }
+      }
+
+      for (let hitbox of hitboxes) {
+        const buffer = _getPositions(this.gltf, hitbox.meshIdx, this.resourceFolder);
+        if (!buffer || buffer.length < 3) {
+          console.warn('Invalid glTF. Could not get buffer.');
+          return undefined;
+        }
+
+        // TODO rotation must be DEFAULT
+
+        const translation = Vector.fromArray(hitbox.node.translation) || Vector.zero;
+        const scale = Vector.fromArray(hitbox.node.scale) || Vector.one;
+
+        let minVector = Vector.fromArray(buffer) as Vector;
+        let maxVector = minVector;
+        for (let i = 0; i < buffer.length / 3; i++) {
+          minVector = minVector.down(Vector.fromArray(buffer, i) as Vector);
+          maxVector = maxVector.up(Vector.fromArray(buffer, i) as Vector);
+        }
+
+        minVector = minVector.mul(scale).add(translation);
+        maxVector = maxVector.mul(scale).add(translation);
+
+        this._hitboxes.push(Box.fromMinMax(hitbox.name, minVector, maxVector));
+      }
+    }
+
+    return this._hitboxes;
   }
 }
 
 function _toRotation(q: { w: number, x: number, y: number, z: number }) {
   const acos = 2 * Math.acos(q.w);
   return q.y > 0 ? Math.PI * 2 - acos : acos;
+}
+
+function _findNodes(gltf: any, name: string, resourceFolder: string, all: boolean = true) {
+  const result = [];
+
+  let nodeIdx = -1;
+  let meshIdx = -1;
+  for (let idx = 0; idx < gltf.nodes.length; idx++) {
+    const node = gltf.nodes[idx];
+    if (all && (node.name?.startsWith(name) || gltf.meshes[node.mesh]?.name?.startsWith(name)) ||
+        !all && (node.name === name || gltf.meshes[node.mesh]?.name === name)) {
+      nodeIdx = idx;
+      meshIdx = node.mesh;
+
+      const buffer = _getPositions(gltf, meshIdx, resourceFolder);
+      if (!buffer) {
+        console.warn(`Invalid glTF. Buffer for node ${nodeIdx} not found.`);
+        continue;
+      }
+      const indices = _getIndices(gltf, meshIdx, resourceFolder);
+      result.push({
+        nodeIdx,
+        meshIdx,
+        name: node.name,
+        translation: Vector.fromArray(gltf.nodes[nodeIdx].translation),
+        scale: Vector.fromArray(gltf.nodes[nodeIdx].scale),
+        rotation: Quaternion.fromArray(gltf.nodes[nodeIdx].rotation),
+        buffer: buffer as ArrayLike<number>,
+        indices: indices
+      });
+
+      if (!all) {
+        return result;
+      }
+    }
+  }
+  return result;
+}
+
+function _findFirstNode(gltf: any, name: string, resourceFolder: string) {
+  const nodes = _findNodes(gltf, name, resourceFolder, false);
+  if (!nodes || nodes.length < 1) {
+    return undefined;
+  }
+  return nodes[0];
+}
+
+function _getPositions(gltf: any, meshIdx: number, resourceFolder: string) {
+  const accessorIdx = gltf.meshes[meshIdx].primitives[0].attributes.POSITION;
+  const ACESSOR_TYPE_VEC3 = 3;
+  return _getBuffer(gltf, accessorIdx, resourceFolder, ACESSOR_TYPE_VEC3, Float32Array);
+}
+
+function _getIndices(gltf: any, meshIdx: number, resourceFolder: string) {
+  const accessorIdx = gltf.meshes[meshIdx].primitives[0].indices;
+  if (accessorIdx === undefined) {
+    return undefined;
+  }
+  const ACESSOR_TYPE_SCALAR = 1;
+  return _getBuffer(gltf, accessorIdx, resourceFolder, ACESSOR_TYPE_SCALAR, Uint16Array);
+}
+
+function _getBuffer(gltf: any, accessorIdx: number, resourceFolder: string, numComponents: number, typedArray: any) {
+  const accessor = gltf.accessors[accessorIdx];
+  const bufferView = gltf.bufferViews[accessor.bufferView];
+  const bufferInfo = gltf.buffers[bufferView.buffer];
+
+  const bufferFile = path.join(resourceFolder, bufferInfo.uri);
+  const buffer = getBuffer(gltf, bufferView.buffer, bufferFile);
+  if (buffer) {
+    const bufferOffset = bufferView.byteOffset || 0;
+    const bufferLength = bufferView.byteLength;
+    const bufferStride = bufferView.byteStride;
+    const bufferViewBuf = buffer.subarray(bufferOffset, bufferOffset + bufferLength);
+    const accessorByteOffset = accessor.byteOffset || 0;
+
+    return buildArrayBuffer(typedArray, bufferViewBuf, accessorByteOffset, accessor.count, numComponents, bufferStride);
+  }
+
+  return undefined;
+}
+
+function _readVectors(node: { translation: Vector | undefined, scale: Vector | undefined, buffer: ArrayLike<number> } | undefined) {
+  if (!node) {
+    return [];
+  }
+  const translation = node.translation || Vector.zero;
+  const scale = node.scale || Vector.one;
+
+  const result = [];
+  for (let i = 0; i < node.buffer.length / 3; i++) {
+    const v = Vector.fromArray(node.buffer, i);
+    if (v) {
+      result.push(v.mul(scale).add(translation));
+    }
+  }
+
+  return result;
+}
+
+/** quick hack sort. take first 2 triangles (first 4 positions) and swap the last two if needed */
+function _sortVectorsXZ(vectors: Vector[]) {
+  if (vectors.length < 4) {
+    return [];
+  }
+
+  if (vectors[1].x.toFixed(4) === vectors[2].x.toFixed(4) || vectors[1].z.toFixed(4) === vectors[2].z.toFixed(4)) {
+    return vectors;
+  }
+
+  return [ vectors[0], vectors[1], vectors[3], vectors[2], ...vectors.slice(4) ];
 }